requirements.txt
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==3.0.1
distlib==0.3.6
filelock==3.9.0
idna==3.4
ischedule==1.2.6
platformdirs==2.6.2
pyTelegramBotAPI==4.9.0
python-dotenv==0.21.1
requests==2.28.2
soupsieve==2.3.2.post1
urllib3==1.26.14
virtualenv==20.17.1
